requirements.client.txt
# client side requirements

numpy
pyzmq >= 17.1.0  # python zmq